Move LoadingSvg to src/assets directory since it will be widely reused
