+++ frontend/src/features/replies/components/ReplyAuthorThumbnail.jsx
@@ -1,6 +1,6 @@
 import { Suspense, useState } from "react";
 import PostAuthorDropdown from "../../posts/components/PostAuthorDropdown";
-import LoadingSvg from "../../posts/assets/LoadingSvg";
+import LoadingSvg from "../../../assets/LoadingSvg";
 import ErrorSvg from "../../posts/assets/ErrorSvg";
 const PostAuthorThumbnail = ({ profile, status }) => {
   const [showDropdown, setShowDropdown] = useState(false);
